requirements.txt
flask
chromadb
torch
torchvision
transformers
pillow